ATR-28: refactoring + folder structure reorganization
