+++ src/Acuminator/Acuminator.Utils/Acuminator.Utils.csproj
@@ -108,13 +108,14 @@
     <Compile Include="RoslynExtensions\PrimaryDacFinder\PrimaryDacTraits\PrimaryDacSpecifiedGraphRule.cs" />
     <Compile Include="RoslynExtensions\PXFieldAttributes\FieldAttributeInfo.cs" />
     <Compile Include="RoslynExtensions\PXFieldAttributes\FieldAttributesRegister.cs" />
+    <Compile Include="RoslynExtensions\SemanticUtils\GraphSymbolUtils.cs" />
     <Compile Include="RoslynExtensions\SyntaxUtils\MethodDeclarationUtils.cs" />
     <Compile Include="RoslynExtensions\SyntaxUtils\SyntaxTreeUtils.cs" />
     <Compile Include="RoslynExtensions\SyntaxUtils\RoslynSyntaxUtils.cs" />
     <Compile Include="RoslynExtensions\BqlModifyingMethods.cs" />
     <Compile Include="RoslynExtensions\PXCodeType.cs" />
-    <Compile Include="RoslynExtensions\ITypeSymbolExtensions.cs" />
-    <Compile Include="RoslynExtensions\CodeResolvingUtils.cs" />
+    <Compile Include="RoslynExtensions\SemanticUtils\ITypeSymbolExtensions.cs" />
+    <Compile Include="RoslynExtensions\SemanticUtils\CodeResolvingUtils.cs" />
     <Compile Include="RoslynExtensions\PXContext.cs" />
     <Compile Include="RoslynExtensions\TypeNames.cs" />
   </ItemGroup>
@@ -126,9 +127,7 @@
     <Analyzer Include="..\..\packages\Microsoft.CodeAnalysis.Analyzers.1.1.0\analyzers\dotnet\cs\Microsoft.CodeAnalysis.Analyzers.dll" />
     <Analyzer Include="..\..\packages\Microsoft.CodeAnalysis.Analyzers.1.1.0\analyzers\dotnet\cs\Microsoft.CodeAnalysis.CSharp.Analyzers.dll" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="RoslynExtensions\SemanticUtils\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="..\..\packages\GitVersionTask.3.6.5\build\portable-net+sl+win+wpa+wp\GitVersionTask.targets" Condition="Exists('..\..\packages\GitVersionTask.3.6.5\build\portable-net+sl+win+wpa+wp\GitVersionTask.targets')" />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
 </Project>
