+++ src/Acuminator/Acuminator.Utils/RoslynExtensions/SemanticUtils/ITypeSymbolExtensions.cs
@@ -154,6 +154,38 @@ namespace Acuminator.Utilities
 			return false;
 		}
 
+		// Determine if "type" inherits from "baseType", ignoring constructed types, optionally including interfaces,
+		// dealing only with original types.
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public static bool InheritsOrImplementsOrEquals(this ITypeSymbol type, string baseTypeName,
+														bool includeInterfaces = true)
+		{
+			if (type == null)
+				return false;
+
+			IEnumerable<ITypeSymbol> baseTypes = includeInterfaces
+				? type.GetBaseTypesAndThis().ConcatStructList(type.AllInterfaces)
+				: type.GetBaseTypesAndThis();
+
+			return baseTypes.Select(typeSymbol => typeSymbol.Name)
+							.Contains(baseTypeName);
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public static bool ImplementsInterface(this ITypeSymbol type, string interfaceName)
+		{
+			if (type == null)
+				return false;
+
+			foreach (var interfaceSymbol in type.AllInterfaces)
+			{
+				if (interfaceSymbol.Name == interfaceName)
+					return true;
+			}
+
+			return false;
+		}
+
 		/// <summary>
 		/// Gets the depth of inheritance between <paramref name="type"/> and its <paramref name="baseType"/>.
 		/// If <paramref name="baseType"/> is not an ancestor of type returns <c>null</c>.
